Explain error when passing compute.googleapis.com to activate_api_identities
Update the description of `active_api_identities` with a note to avoid
passing compute.googleapis.com as an argument. It seems that this API
call is not permitted despite lack of documentation. This can be seen by
calling:

```sh
gcloud alpha services identity create --service=compute.googleapis.com --log-http
```

Calling this same command with, e.g., cloudbuild.googleapis.com does not
result in an error.